+++ manufac-analytics/src/components/footer/Footer.js
@@ -1,0 +1,12 @@
+import React from "react"
+
+function Footer(){
+
+    return(
+        <footer>
+            <h1>Copyright --insert company name-- 2020</h1>
+        </footer>
+    );
+}
+
+export default Footer;
